refactor(ui): migrate Image settings components to TypeScript
