+++ ui/src/app/settings/views/Images/ThirdPartyDrivers/ThirdPartyDrivers.tsx
@@ -1,18 +1,20 @@
-import { Col, Spinner, Row } from "@canonical/react-components";
-import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
 
+import { Col, Spinner, Row } from "@canonical/react-components";
+import { useDispatch, useSelector } from "react-redux";
+
+import ThirdPartyDriversForm from "../ThirdPartyDriversForm";
+
+import { useWindowTitle } from "app/base/hooks";
 import { actions as configActions } from "app/store/config";
 import configSelectors from "app/store/config/selectors";
-import { useWindowTitle } from "app/base/hooks";
-import WindowsForm from "../WindowsForm";
 
-const Windows = () => {
+const ThirdPartyDrivers = (): JSX.Element => {
   const loaded = useSelector(configSelectors.loaded);
   const loading = useSelector(configSelectors.loading);
   const dispatch = useDispatch();
 
-  useWindowTitle("Windows");
+  useWindowTitle("Ubuntu");
 
   useEffect(() => {
     if (!loaded) {
@@ -24,10 +26,10 @@ const Windows = () => {
     <Row>
       <Col size={6}>
         {loading && <Spinner text="Loading..." />}
-        {loaded && <WindowsForm />}
+        {loaded && <ThirdPartyDriversForm />}
       </Col>
     </Row>
   );
 };
 
-export default Windows;
+export default ThirdPartyDrivers;
