+++ ui/src/app/settings/views/Images/VMWare/VMWare.tsx
@@ -1,18 +1,20 @@
-import { Col, Spinner, Row } from "@canonical/react-components";
-import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
 
+import { Col, Spinner, Row } from "@canonical/react-components";
+import { useDispatch, useSelector } from "react-redux";
+
+import VMWareForm from "../VMWareForm";
+
+import { useWindowTitle } from "app/base/hooks";
 import { actions as configActions } from "app/store/config";
 import configSelectors from "app/store/config/selectors";
-import { useWindowTitle } from "app/base/hooks";
-import ThirdPartyDriversForm from "../ThirdPartyDriversForm";
 
-const ThirdPartyDrivers = () => {
+const VMWare = (): JSX.Element => {
   const loaded = useSelector(configSelectors.loaded);
   const loading = useSelector(configSelectors.loading);
   const dispatch = useDispatch();
 
-  useWindowTitle("Ubuntu");
+  useWindowTitle("VMWare");
 
   useEffect(() => {
     if (!loaded) {
@@ -24,10 +26,10 @@ const ThirdPartyDrivers = () => {
     <Row>
       <Col size={6}>
         {loading && <Spinner text="Loading..." />}
-        {loaded && <ThirdPartyDriversForm />}
+        {loaded && <VMWareForm />}
       </Col>
     </Row>
   );
 };
 
-export default ThirdPartyDrivers;
+export default VMWare;
