+++ ui/src/app/settings/views/Images/VMWareForm/VMWareForm.tsx
@@ -1,10 +1,10 @@
 import { useDispatch, useSelector } from "react-redux";
 import * as Yup from "yup";
 
-import { actions as configActions } from "app/store/config";
-import configSelectors from "app/store/config/selectors";
 import FormikField from "app/base/components/FormikField";
 import FormikForm from "app/base/components/FormikForm";
+import { actions as configActions } from "app/store/config";
+import configSelectors from "app/store/config/selectors";
 
 const VMWareSchema = Yup.object().shape({
   vcenter_server: Yup.string(),
@@ -13,7 +13,7 @@ const VMWareSchema = Yup.object().shape({
   vcenter_datacenter: Yup.string(),
 });
 
-const VMWareForm = () => {
+const VMWareForm = (): JSX.Element => {
   const dispatch = useDispatch();
   const updateConfig = configActions.update;
