+++ ui/src/app/settings/views/Images/Windows/Windows.tsx
@@ -1,18 +1,20 @@
-import { Col, Spinner, Row } from "@canonical/react-components";
-import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
 
+import { Col, Spinner, Row } from "@canonical/react-components";
+import { useDispatch, useSelector } from "react-redux";
+
+import WindowsForm from "../WindowsForm";
+
+import { useWindowTitle } from "app/base/hooks";
 import { actions as configActions } from "app/store/config";
 import configSelectors from "app/store/config/selectors";
-import { useWindowTitle } from "app/base/hooks";
-import VMWareForm from "../VMWareForm";
 
-const VMWare = () => {
+const Windows = (): JSX.Element => {
   const loaded = useSelector(configSelectors.loaded);
   const loading = useSelector(configSelectors.loading);
   const dispatch = useDispatch();
 
-  useWindowTitle("VMWare");
+  useWindowTitle("Windows");
 
   useEffect(() => {
     if (!loaded) {
@@ -24,10 +26,10 @@ const VMWare = () => {
     <Row>
       <Col size={6}>
         {loading && <Spinner text="Loading..." />}
-        {loaded && <VMWareForm />}
+        {loaded && <WindowsForm />}
       </Col>
     </Row>
   );
 };
 
-export default VMWare;
+export default Windows;
